requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
reportlab==4.0.7
matplotlib==3.8.2
pillow==10.1.0
